creación de la vista de listar inmuebles
